add user system with breez

+++ app/Http/Controllers/TasksController.php
@@ -15,13 +15,21 @@ class TasksController extends Controller
      */
     public function index()
     {
-        // メッセージ一覧を取得
-        $tasks = Task::all();         // 追加
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            // （後のChapterで他ユーザの投稿も取得するように変更しますが、現時点ではこのユーザの投稿のみ取得します）
+            $tasks = Task::orderBy('created_at', 'desc')->paginate(10);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            }
 
-        // メッセージ一覧ビューでそれを表示
-        return view('tasks.index', [     // 追加
-            'tasks' => $tasks,        // 追加
-        ]);                                 // 追加
+        // dashboardビューでそれらを表示
+        return view('dashboard', $data);                               // 追加
     }
 
     /**
@@ -47,23 +55,22 @@ class TasksController extends Controller
      */
     public function store(Request $request)
     {
-
         // バリデーション
         $request->validate([
             'title' => 'required|max:30',   // 追加
             'content' => 'required|max:255',
             'status' => 'required|max:10',
-        ]);        
-
-        // メッセージを作成
-        $task = new Task;
-        $task->title = $request->title;
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
-
-        // トップページへリダイレクトさせる
-        return redirect('/');
+        ]);
+        
+        // 認証済みユーザ（閲覧者）の投稿として作成（リクエストされた値をもとに作成）
+        $request->user()->tasks()->create([
+            'title' => $request->title,
+            'content' => $request->content,
+            'status' => $request->status,
+        ]);
+        
+        // 前のURLへリダイレクトさせる
+        return back();
     }
 
     /**
@@ -109,7 +116,9 @@ class TasksController extends Controller
      */
     public function update(Request $request, $id)
     {
-        
+        // idの値でタスクを検索して取得
+        $task = Task::findOrFail($id);
+
         // バリデーション
         $request->validate([
             'title' => 'required|max:30',   // 追加
@@ -117,16 +126,14 @@ class TasksController extends Controller
             'status' => 'required|max:10',
         ]);
         
-        // idの値でメッセージを検索して取得
-        $task = Task::findOrFail($id);
-        // メッセージを更新
-        $task->title = $request->title;
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
+        $task->update([
+            'title' => $request->title,
+            'content' => $request->content,
+            'status' => $request->status,
+        ]);
 
-        // トップページへリダイレクトさせる
-        return redirect('/');
+        // 前のURLへリダイレクトさせる
+        return back();
     }
 
     /**
@@ -137,12 +144,18 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        // idの値でメッセージを検索して取得
-        $task = Task::findOrFail($id);
-        // メッセージを削除
-        $task->delete();
+        // idの値で投稿を検索して取得
+        $task = \App\Models\Task::findOrFail($id);
+        
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+            return back()
+                ->with('success','Delete Successful');
 
         // トップページへリダイレクトさせる
-        return redirect('/');
+        return back()
+            ->with('Delete Failed');
+        }
     }
 }
